requirements.txt
fastapi
uvicorn
httpx
black
python-dotenv
loguru
